deprecate AlgorithmsForKey, move to internal
