+++ agents/docs/packages.md
@@ -62,9 +62,11 @@ JSON Web Signatures per RFC 7515. Sign, verify, parse.
 - Key types: `Message`, `Signature`, `Headers`, `KeyProvider`, `KeySink`, `Base64Encoder`
 - Options: `WithKey()`, `WithKeySet()`, `WithVerifyAuto()`, `WithJSON()`, `WithDetachedPayload()`, `WithDetachedPayloadReader()` (streaming variant; single-key, HMAC/RSA/ECDSA only)
 - Global/per-call settings: `WithMaxSignatures()` (usable in both `Settings()` and `Parse()`/`ReadFile()`)
-- Registration: `RegisterSigner()`, `RegisterVerifier()`, `AlgorithmsForKey()`, `RegisterAlgorithmForKeyType()`, `RegisterAlgorithmForCurve()`
+- Registration: `RegisterSigner()`, `RegisterVerifier()`, `RegisterAlgorithmForKeyType()`, `RegisterAlgorithmForCurve()`
+- **AlgorithmsForKey()** — DEPRECATED, do not use and do not add callers. It was never meant to be consumed by end users; it is jwx's internal algorithm-inference helper, exported for historical reasons only. The implementation lives in `jws/internal/keyalg` (`keyalg.Candidates`), which every in-tree caller uses directly. It is not a key/algorithm compatibility check, and its output may widen or narrow in any release.
 - Error sentinels: `SignError()`, `VerifyError()`, `VerificationError()`, `ParseError()`
 - Sub-package: `jws/jwsbb` — compact serialization, signing, verification building blocks. The fastjson-backed `Header` probe (`HeaderParse`/`HeaderGet*`/`HeaderHas`) is a thin facade over `jws/internal/jwsbb`, which holds the implementation plus jwx-internal-only helpers not re-exported by the facade (e.g. `HeaderForEachKey`, the header-key enumerator used by `VerifyCompactFast`'s minimal-shape gate)
+- Sub-package: `jws/internal/keyalg` — key-to-algorithm classification. Owns the process-global registration tables behind `RegisterAlgorithmForKeyType`/`RegisterAlgorithmForCurve` (as `RegisterForKeyType`/`RegisterForCurve`), plus `Candidates` (which algorithms a key could be used with), `KeyTypesFor` (reverse lookup), `ValidateEd25519KeyShape`, and the `ErrUnclassifiableKey` sentinel. Registers the builtin algorithms from its own `init()`.
 - ML-DSA (go1.27): `mldsa.go` installs a `Signer`/`Verifier` pair per parameter set. Both accept a raw `crypto/mldsa` key or an AKP `jwk.Key`, and both reject a key whose parameter set disagrees with the algorithm. The signature primitives come from `lestrrat-go/dsig` v1.4.0, which owns the algorithm registration.
 - Files: `jws.go`, `message.go`, `signer.go`, `verifier.go`, `headers.go`, `interface.go`, `errors.go`, `options.go`, `key_provider.go`, `sign_context.go`, `verify_context.go`, `streaming_detached.go`, `mldsa.go` (go1.27)
 - Imports: jwa, jwk, cert, dsig, internal/{base64,json,pool,tokens}
